fix: when redefining var, use matching pattern name
Given the following program:

```
let me|I be false

to renew——:
    & true.

let me be renew——
me
```

The final line would print `false`, the original value of the variable. This was happening because when redefining the variable on line 6, it would look up `me` in the set of name/value pairs in the environment. It found the existing variable, which had the RegEx pattern `me|I`, but when setting the new value, it was using the given name instead of the existing pattern. This caused a new entry to be entered into the set, so that both `me|I` (with the value `false`) and `me` (with the value `true`) existed.

This change uses the existing pattern in the set to update the new value for the variable.